+++ core/nodes.py
@@ -2,7 +2,7 @@ import os
 
 from PyQt5.QtGui import QStandardItem
 
-from pkg import file
+from core import file
 from gui import icons
 
 TreeNode = list[QStandardItem]      # Type alias for return values
