+++ tree.py
@@ -5,8 +5,7 @@ from PyQt5 import QtCore
 from PyQt5.QtCore import QRunnable, pyqtSignal, QObject
 from PyQt5.QtGui import QStandardItemModel
 
-from pkg import file
-from core import nodes
+from core import nodes, file
 
 
 class TreeType(Enum):
